requirements (1).txt
gradio
torch
transformers
sentence-transformers
faiss-cpu
PyMuPDF
python-dotenv
accelerate